Added cost function mask.
Added cost function mask variable to Input_Opt to track where
cost function should be evaluated.

At present, ExtData is not able to read in cube-sphere files
so to evaluate the cost function on the cube-sphere, I have added
ranges for I,J and L in GCHP.rc to define the region.  ExtData
should be able to read CS input files in GCHP v 13. Also at present
the entire CF region must be on a single PET. This is obviously
not ideal but only requires some changes in the loop in
gigc_chunk_init where the mask is filled in.

Fixed bug which was causing adjoint to crash after 979 (or so)
chemistry timesteps because MAPL_DestroyFile doesn't actually close
the file or free it's LUN. Added call to FREE_FILE which does that
so we no longer run out of file handles.

Changed a loop name in FVdycore because defining ADJOINT for
preprocessor ifdefs was filling in a 1 there and breaking FVdycore
compiling. Renamed loop from ADJOINT to ADJOINT_LOOP.